+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 3.42.22
+
+### Changed
+
+- Demo Data update
+- Event handlers do use their own service id now
+- Update ti-business-partner-identity-activated-event to 3.0.0.
+- Update ti-business-partner-identity-updated-event to 3.0.0.
+
+## 3.42.21
+
+### Changed
+
+- fixed race condition when publishing vqps submissions (cloud-native approach by using db polling)
+
 ## 3.42.20
 
 ### Changed